Created carousel component for testimonials
